kafka_clients/requirements.txt
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.2
joblib==1.3.2
EOF